Rename file containing interface QueryExecutionResponse.
