+++ server/interfaces/QueryExecutionResponse.ts
